Break down lambda functions into several files
